Change Image Size to be smaller For some reason next/image is not working for me locally. I'll have to investigate further what's happening. In general though, from troubleshooting, I am wholly unimpressed by next/image and the many constraints with image styling.
In the meantime, to reduce loading times and prevent large downloads,
the first image is now significantly smaller.

This commit also includes some minor capitalization fixes and centering
the title. Applying `text-center` to a `<p>` will center it within the
paragraph, but makes no guarantees the `<p>` itself will be centered.